Convert test_error_in_processing_tx() to a mempool test
